==> 20_reproducibility.qmd <==
---
title: '20: Reproducibility'
jupyter:
  jupytext:
    text_representation:
      extension: .qmd
      format_name: quarto
      format_version: '1.0'
      jupytext_version: 1.16.1
  kernelspec:
    display_name: Python 3 (ipykernel)
    language: python
    name: python3
---

<https://bellecp.github.io/597-Data-Wrangling-Spring-2024/project.html#provide-instructions-on-how-to-reproduce-the-project>.
Describe your contributions below.

* * *

### Requirements for this project

Once the dependencies are installed, you can proceed to run the code and reproduce the results as described in the project documentation

```{python}
#pip install -r requirements.txt
```

### Rquired Libraries 

'''
matplotlib==3.5.1
seaborn==0.11.2
pymongo==4.0.0
numpy==1.21.4
pyxirr==0.10.3
pandas==2.2.1
pymongo==4.6.3
plotly==5.9.0
matplotlib==3.7.1
numpy==1.26.3
seaborn==0.12.2
jupyter==1.0.0
pytest==7.1.2
ipywidgets==8.0.4
ipython==8.15.0
scikit-learn==1.2.2
statsmodels==0.14.0
'''

```{python}
# pip list
```
